[docs] Add troubleshooting for GUI issue
miraclGUI does not work when executed in a Docker container locally or
in a Singularity container on a SLURM cluster. There are several
troubleshooting steps that need to be added to the docs. In this commit,
steps have been added to the docker installation doc. WIP at this point,
needs to be completed.