[docs] update MRI reg tutorial from Ants to nifty
Change 'MRI whole-brain registration to Allen atlas tutorial' from
'mri_allen_ants' to 'mri_allen_nifty' to reflect function calls in
MIRACL cli and GUI. Also added screenshot of registration options
window to GUI part of the tutorial as it was referenced in-text but not
actually included. Lastly, added missing option for GUI and deleted GUI
option that is only available in the cli.