[fix!] upgrade base image container to Focal 20.04
The Neurodebian version of the Docker base image was Xenial 16.04. The
main reason for updating to 20.04 was the installed version of glibc
which was too old to work correctly with components of the STA workflow,
causing the latter to crash. Updating to 20.04 fixes this problem.
However, the update to 20.04 will break the base image Dockerfile as
some of the software that is being installed (e.g. FSL) requires a
different installation procedure under 20.04.

[BREAKING CHANGE] Base Dockerfile not compiling correctly due to update
to 20.04.