[feat] add cust build script and docker-compose
Adds a custom build script that builds the Docker image from the
Dockerfile in the same directory. The build script adds the user as well
as the user and group id as environmental arguments required in the
Dockerfile. This will allow the image to be build with the host user as
the Docker container user. This commit also includes the addition of a
docker-compose.yml which is used to start the Docker container from the
image built with the build script. The docker-compose.yml imports the
host $DISPLAY into the container as well as specifying the Docker
network used for the container which in this case is the host network.
The docker-compose.yml also mounts a volume into the container which
includes the .Xauthority file. The correct path for the volume is set by
the build file using regex. $DISPLAY, host network and .Xauthority are
required to make X11 work with MIRACL in Docker.