[feat] implement NiftyReg installation
Implements NiftyReg installation into Dockerfile. NiftyReg gets installed
with the default make parameters. They can be changed manually before
creating the MIRACL image with the build script. In a future integartion
the build script should allow arguments to be passed for the make parameters.
This commit officially adds the installation of NiftyReg in the MIRACL
Docker environment as default.