[docs] update Singularity installation docs
Updates installation and usage docs with new pull method. Instead of
building the Singularity container by pulling it from Docker Hub, a
pre-built container of the latest version of MIRACL can now be
conveniently be pulled from the Sylabs cloud. The old method of pulling
a container from Docker Hub is still included but has been moved to the
Troubleshooting section of this doc.