[docs] update Docker installation docs
Updates installation and usage instructions for using MIRACL with Docker.
The updated method for building Docker is using the build script we
provide. Images now have to be run using Docker Compose. This is to
prevent issues with the MIRACL GUI and X11 forwarding in Docker. The
previous way of using MIRACL and Docker is still included but has been
moved to the Troubleshooting section of the Docker installation docs.