[fix] compile NiftyReg with g++-5
NiftyReg compilation would fail using g++-9 which is installed in
Neurodebian Focal. Using g++-5 instead fixes this issue. g++-5 is
installed in the base Docker image and pointed to before
compiling NiftyReg. After compiling is done g++ is pointed back to
version 9. This commit also fixes a pip resolver issue that checked for
compatible versions of botocore. The resolver would download each version
which took a lot of time and bandwidth to finish. Issue got fixed by
explicitely telling the resolver which versions to look for (>=
1.20.110). Lastly, Docker pulls the image from a new Docker Hub location
that references the revised base Docker image.