[feat] add Singularity to CircleCI
Adds complete Singularity CircleCI workflow. Uses quay.io/singularity
container (v3.10.3) for continuous integration. First, spython is
installed through pip. This is needed to convert the existing Dockerfile
to a Singularity def file. Next, permissions for the /code directory are
changed recursively in the Dockerfile to allow the (non-root)
Singularity user access to it. The next step is to log in to the Sylabs
cloud where the compiled Singularity image will be stored. The password
is being read from a secure CircleCI environment file (context). Next,
the current image on the Sylabs cloud is being deleted. This is
unfortunately required due to storage restrictions in the cloud. If no
image can be found, the returned error will be caught by a trap. Lastly,
the Singularity container is build in a remote Singularity environment
outside of the current CircleCI Docker environment and automatically
pushed to the cloud. This entire process will only happen if a commit is
being pushed to main and ignored for other branches.