[docs] change X11 issue troubleshooting entry
The entry for making the MIRACL GUI work using Singularity and X11 on a
cluster is convoluted and unlikely to work. Instead users should just
use VNC. The instructions have been updated accordingly including a
recommendation for the latter. The old troubleshooting instructions have
been kept, however, in case a user has a reason not to use VNC and needs
to make the MIRACL GUI work with X11 directly. These instructions might
be removed in a future version of MIRACL.