[fix] add unrecognized flags for STA flow in CLI
Several flags were not recognized when running the STA workflow from the
CLI. This adds the flags that are required at a minimum to run the STA
workflow successfully. There are several flags that have not yet been
added and can therefore not be used when running the STA workflow from
the cli. Also fixes minor mistakes: 1) niftis directory was hardcoded,
2) flags overlapped, 3) step length logic was incorrect (inverted), 4)
downsample ratio was parsed incorrectly (0 needed to be prepended for
single digit downsample values).