[fix] correct label formatting for split hemi
Allen label needs to be provided with R or L prefix for hemisphere when
chosing hemi split. The prefix will get cut before the depth is looked
up. However, if the label is not provided with a prefix but split is
used the label will still be cut and the process will fail. This commit
fixes that by using the uncut label if the cut label does not exist in
the atlas. If neither the cut nor the original label exist, the process
will fail with an error message and exit status 1.