[fix] check for latest Docker Compose version
Adds additional check for latest version of Docker Compose. Docker
Compose can be installed as a standalone on Linux, as a plugin and as
part of Docker Desktop. If the standalone is not found, the build.sh
script will check for the plugin and Docker Desktop versions and return
a custom message if found.