[fix] disable range-kutta method parsing
The range-kutta method is currently not working. I am not entirely sure
why yet but the variable rk2 gets passed to the miracl sta track_tensor
function with the -r flag which is not present in this function. This
results in an error. The temporary fix here is to assign an empty string
to the rk2 variable that gets evaluated correctly by the if expression
that tests whether the range-kutta method should be used. An empty
string results in not using the method.