[fix] install setuptools v65.6.0
setuptools v67.0.0 introduced changes that broke the MIRACL CLI and GUI
methods. Installing version v65.6.0, which has previously been tested
with MIRACL, fixes this issue. However, this should only be a temporary
fix. Ideally, the affected scripts should be changed in MIRACL to make
them work with v67.0.0 of setuptools. The reason why this is not being
done in this commit is, as so often, time constraints.