[docs] squash start of refactoring docs
[docs] fix spelling mistake in doc folder and file

[docs] add missing batch arg to Singularity cmd

[docs] add missing batch arg to Singularity cmds

[docs] create folder for MIRACL's menu images

Created folder for all images of MIRACL's menus. For example MIRACL's
main menu or the main registration menu. From now on, all tutorials that
use images of MIRACL's menus should link to here to avoid that the same
images have to be changed in several places when being updated.

[docs] update to correct miracl call in cli

The miracl command that was used in the example was not up to date
anymore. Changed to correct command.

[docs] correct spelling and formatting mistakes

[docs] add link to tutorial section

[docs] update link to MIRACL menu image

Changed link to the updated image for MIRACL's menu at the updated
location: gallery/menu

[docs] update link for MIRACL's main menu

Also changed description for image.

[docs] change descriptions for gallery images

The descriptions for the gallery images were the image names which did
not make sense. Changed the descriptions to make sense and deleted them
where they were not needed.