[docs] squash restructuring/improvements/fixes
[docs] test different formatting for docs markdown

Test to see if a different format works better for readability of
function options.

[docs] change markdown format for more testing

[docs] change formatting back to original version

Formatting tests were not successful. The original formatting using
bullet points seem to work best for the GitHub flavor of Markdown.

[docs] improve formatting

[docs] fix formatting

[docs] test different formatting

[docs] test new about docs structure

This is testing the different structure for MIRACL's docs. The about
section now encompasses more topics that are separated and indexed in
the about.md file. This will be reverted if it doesn't work as intended,
which is to make the docs more structured and readable.

[docs] add references and format section

Adds correct APA and BibTeX references. Also improves formatting.

[docs] correct BibTex formatting

[docs] add about section to docs index

[docs] add link to GPLv3 license

[docs] delete introduction and link to index

[docs] format index

[docs] change order and add link to index

[docs] change e-mail link format

[docs] change e-mail link format

[docs] add link for HippMapp3r

[docs] move installation files to separate folder

Moved all installation md files in a separate folder and improved their
respective formatting. The latter can be reverted if the new structure
is not improving readability of the docs.

[docs] delete installation files after move

[docs] change installation instructions

Renamed and changed the formatting of installation methods overview file.

[docs] improve user guide

Renamed modules and workflows overview to user guide and improved
formatting of user guide. Also added examples of how to use modules.

[docs] change formatting

[docs] change title

[docs] update index

Updated the index file to reflect docs changes. Also temporarily deleted
the full-documentation entry as the pdf is currently not rendering
correctly.

[docs] change title and formatting

[docs] delete after renaming

[docs] change order of topics

[docs] complete restructuring of About section

About section needed to be restructured due to the limitations of
Sphinx.

[docs] delete separate About section files

Unfortunately, due to the limitations of Sphinx, the about section
needed to be restructured and the neatly organizaed, separate About
section files were deleted.

[docs] restructure installation docs

Because of Sphinx limitations, all installation instructions needed to
be combined in the same document to unite them under one header in the
sidebar.

[docs] custom toc title

[docs] improve formatting

[docs] add papers citing MIRACL to Citations

These are only some papers. More publications need to be added in the
future. WIP!

[docs] improve formatting

[docs] add license image to readthedocs index

[docs] test fix to link to page from index.rst

[docs] test fix to link to page from index.rst

[docs] test fix to link to page from index.rst

[docs] fix links to .md page from index.rst

[docs] fix spelling mistake

[docs] fix spelling mistake

[docs] fix grammar