[fix] change detection method for MIRACL version
The build script did not correctly detect the version for the Docker
image. It would parse the version.txt file and assign that version to
the Docker image even if the latest image was pulled. This has been
fixed by adding an additional flag '-s' to run when building an image
from a different MIRACL version. The version of the build script hasn't
been incremented yet as the updated script has not been fully tested
yet. Once it has been test, the version number will be incremented.