[docs] squash corrections/improvements/versions
[docs] increment version of build script

[docs] correct docs for using specific version

Downloading a specific version older than MIRACL >= 2.2.4 does not
include the latest version of the build script. The user has to either
build the older version using its respecitve instructions or download
the latest version of the build script and use it with the older version
to build the image.

[docs] squash troubleshooting sections into one

The separate links to the troubleshooting section of the docs have all
been combined to one single link at the end of the installation
instructions. Also made small changes to layout by introducing
separators between sections.

[docs] add content for about section

[docs] improve formatting