[docs] correct docs for using specific version
Downloading a specific version older than MIRACL >= 2.2.4 does not
include the latest version of the build script. The user has to either
build the older version using its respecitve instructions or download
the latest version of the build script and use it with the older version
to build the image.