[docs] replace md docs with rst docs - conf files
Updating general configuration files to migrate from Sphinx docs written
in Markdown to the more powerful rst format. This new version of the
documentation is completely written in rst format. A requirements.txt
has been added that includes the extentions that must be installed in
order for the improved docs to work. Also changed the custom css
implementation from the deprecated, current implementation to the latest
recommended implementation. The extentions that are required are being
installed by including them in setup.py but this should be done
differently in the long run. It's done this way now for testing
purposes.