[fix] fix parse_inputs fn
The parse_inputs function did not include the added arguments vx and vz.
It subsequently also didn't return them when being unpacked in the main
function. Arguments have been included and are being returned correctly.
A more detailed output for the user to stdout of which arguments are
being used for voxelization has also been added.