[fix] correct path to Allen Atlas and improve fn
The environmental variable which is supposed to be the path to the Allen
Atlas was incorrect. Added a new environmental variable to the
Dockerfile that is used in the script instead to fix the filepath issue.
Also added a new flag to the voxelization function that let's the user
choose between 10 and 25 resolution.