[fix] pin SimpleITK to version 2.2.1
The latest version of SimpleITK, v2.3.1, requires wheels that are build
when the Docker container is being built. This process takes >20mins
which will break the CI pipeline of CircleCI. Pinning the Version of
SimpleITK to v2.2.1 fixes this as this previous version doesn't require
wheels. However, this should be considered a temporary fix. There are
two steps that have to be taken to make this solution more permanent: 1)
set up the requirements.txt file with setup.py to pin all packages with
their respective dependencies without having to overcrowd setup.py and
2) test downloading the wheels for SimpleITK manually and installing
them with pip if the latest version of SimpleITK is required.