[fix] fix incorrect url from readthedocs to GitHub
Before applying this fix, clicking the link on readthedocs to edit a page on the GitHub repo would yield a page not found error. The /docs folder would be replicated in the url. Removing it in de conf.py file fixed this error.
The error, btw, was caused after adding the .readthedocs.yaml to the source directory as required by readthedocs in their new guidelines from ~September 2023.