[docs] added category, changed parts of structure
The data category has been replace with a downloads category. The
content of data is now a subcategory of downloads. Another subcategory
called workshops was added to the downloads header. The workshops
subcategory contains the material for workshops that were given on
MIRACL. This particular update contains the notebooks for the Stanford
workshop. They are currently in a data folder that is hosted directly on
GitHub but all data should eventually be moved to a different,
centralized location. The notebooks in this update also still need to be
cross-referenced with links to their respective tutorials.