[feat] first version of Fischer atl implementation
First iteration of Fischer Reference atlas for rat brains implementation
into mri reg. The name of the fn should be changed to reflect the
changes i.e. it should be changed to a generic 'atlases' or
'allen_fischer' or sth.  Also, parsing logic should be tested and
default should be defined when e.g. custom lbls or resolution args are
provided.