[feat] add first version of ACE flow GUI
This is the first version of the ACE flow GUI. The user input has not yet been connected to the cli logic.
The GUI is implemented using a Model-View-Controller architecture as well as a tab manager and a widget
utility fn. The View for the main ACE tab is complete. The user can toggle between single and multiple method.
The user can also show/hide the tabs for the modules that are being used by the ACE workflows. The main ACE
tab contains the required and most important/useful cmds to run ACE. The scripts for voxelization/warping and
clarity registration are already present but have to be populated. Once all modules have been integrated, the
logic needs to be connected.