[feat] add script to extract cli parser help text
Working on a script that parses the cli argparser file for ACE and extracts flag/help text pairs
into a Python dictionary. The help texts from the dictionary can then be referenced in the label
widgets to populate the pop up help text when the mouse pointer is hovered over a label. This script
removes '\n' characters and correctly replaces placeholders with the default for the parser args.