[feat] replace shm_size with mem_limit
Instead of using the -d flag to set the memory size that is shared between containers,
the flag now sets the max memory limit for the container itself. For that the script
first checks whether the host machine is running Linux or MacOS and uses the appropriate
command to set 85% of available memory as a default.