[feat] implement label tool tip from parser help
Created a utility function for label widgets that parses the ACE parser
and returns a dict of key=flag and value=help. The text from the parser
help is used as the floating tool tip for the label. The idea is that
when the help text is changed in the parser, the changes will propagate
to the floating tool tip without having to change the floating tool tip
manually.