[fix] correct default ref for --rca_skip_cor
Testing different methods for creating the miracl flow ace command with
flags based on user input. Sever utility methods in this commit will
probably become obsolete with the next iteration. The logic for creating
the flags should be moved to a separate script. However, the logic flow
seems to be working and is relatively robust. A dict is created with
keys=verbose_flags and values=user_input. The MIRACL command is created
based on certain conditions for each flag etc.