[feat] complete vox/warp and add/amend util widgets
Voxelization/warping is now implemented and linked to logic. Also
amended the fn to choose a folder. It's now a choose path fn that let's
the user decided whether to select the path to a folder or a file. Also
added a new utility widget that creates n text fields with n
corresponding labels.