[feat] add correlation/stats tab and amend text field widget
Correlation/stats tab is now implemented. Decided to combine correlation
and stats because having two different tabs for so few widgets would be
wasteful. Also amended the text field util widget. It now has an
additional regex validator that is alphanumeric but also allows "-" and
"_" to connect strings as spaces are not allowed.