[fix] add workaround for reg module boolean flags
Some of the flags in the CLARITY registration module have no input.
Rather they are set as booleans. This required a workaround in the flag
creator as well as in the flag dictionary creator. In the process I
discovered an incorrect default value of one of the flags in question.