[fix/wip] fix default value not passed in three text widget creator method
The create_three_text_input_widget method did not automatically return
the default value if the text widget is empty i.e. there was no user
input, only the placeholder. This was fixed in the flag creator by using
'or' for 'var or default'.

Also started to refactor widget creation with the new argparser
extraction method.