[feat/wip] add flag config and new method to widget utils
The way the flags are currently used is not great since there is too
much duplication going on. What should happen is that there is an object
that includes the flags and all other options from the parser. This
object should get the flag mostly dynamically so that if a flag gets
changed in the parser, it will only have to be updated in the flag
config and nowhere else.