[feat] separate module from flow flags
I realized that the base flags for the modules should not be the flags
that are being used in the workflow. I added a new attribute to the
datamodel called 'flow' that can contain dictionaries for workflows.
E.g. 'ace' will contain a short and a long flag for the module argument
that is used in the workflow. The base short and long flags will be the
actual flags from the module itself. I also implemented checks in the
parser creator function to deal with the different cases e.g. if no
workflow flag is present (if the argument is part of the workflow but
not part of a separate module the base flag would be the workflow flag)
the base flag is used.