[wip/feat] add reset/load/save functionality
First implementation of reset/load/save functionality. Separated this
functionality as a service pattern into 'miracl_gui_data_manager'. Also
streamlined the tab controller to only export the dictionaries that are
actually needed for the functionality. This gets rid of duplication.
Also added an observer pattern to detect changes in the widgets in real
time to allow for data input validation. The observer is
'miracl_gui_widget_observer'.