[wip/fix/feat] add multiple widgets functionality and update logger
This took way too much testing to implement but it finally seems to be
working. QSpinBoxes spawn underlying QLineEdit widgets and process their
values like a wrapper using the value() call. This caused problems when
iterating over the children of a container (object containing multiple
widgets with the 'inputs' attr). I fixed it by checking for widget names
duplications and it's now working for QLineEdit and QSpinBox widgets. If
the default attr of the MiraclObj is a single value (str, int, float),
one widget will be created for it. However, if the attr is a list, the
function will create n widgets where n is the length of the list i.e.
number of defaults. The widgets will get created on the same line
spanning the same width as a single widget. The values from each widget
are picked up by the observer for detecting real time changes and will
be returned as a list in the values dictionary. That way they will be
correctly parsed by the actual module fn call. To debug this I used my
default DAP setup but also modified the logger. It is now able to create
multiple instances that can be used to log specific parts of the script
while not logging others. All logging can still be disabled globally but
this makes for a much cleaner debugging experience.