[wip/fix] start migrating from mri allen nifty reg to mri nifty reg
The idea is to replace the current Allen registration with a script that
allows the user to use different atlases (Allen, Fischer etc.). Copied a
version of the new script over from the Fischer implementation branch.
This needs to be tested before inplmenting further atlases. Also, the
inversion of affine and cpp needs to be discussed for the new
implementation.