[feat] add new method of including atlases in Dockerfile
This method involves hosting the atlases on Huggingface and using their
cli application to download them.

[fix] fix spelling mistake in target directory

[fix] change atlas download strategy

It would be much nicer to use the Hugging Face CLI but unfortunately
that's not possible with the current version of MIRACL. Until we get the
big refactor done we will have to download the atlases using 'git
clone', copy them them in the '/code' folder and remove the downloaded
repo. Doing it like this at least avoids having to tar.gz them in the HF
repo. That way we don't have to remember to create a new tar.gz version
of the atlases when the content of the atlas directory changes.

[wip] do not download HuggingFace cli

[fix] enable lfs for git to correctly download files from HF repo