[docs/feat] update docs on MRI Niftyreg registration module
Temporarily deleting the GUI entry for the module as it needs to be
refactored. It is currently not working. The question is whether we
should fix it now in its Bash script version or wait for the Pydantic
object architecture to be ready and then refactor the entire script. I
think the latter is the way to go.