[docs] update tested OS list and add warning about sudo use with install.sh
Some users have to run Docker with "sudo" as they don't have a "docker"
group. This will cause problems with the install script as it will try
to create a user "root" inside of the container. "root" being the
default Docker user, this will fail. Users are therefore warned that the
installation script can only be used if Docker can be used without
"sudo" on their system. The next step here would be to add a check in
the installation script that throws an error when the detected uid is
"root".