[feat] add interactive prompt to install.sh
The interactive prompt will guide the user through the most important
arguments that have to be provided to the installation script. The
prompt is an abbreviated version of the cli flags. The reasoning behind
that is that users who need to use the prompt instead of providing the
flags are probably not advanced enought to use the extra flags in the
first place.