[feat] add random name generator fn
One problem that users ran into was duplicated image/container names.
Sometimes they would install MIRACL but then use the install.sh script
to install MIRACL again without first removing the previous Docker
image/container. This would lead to build conflicts. This could be
solved if users used the -i and -c flags to generate custom names.
However, it can't be assumed that user do that so now instead of a hard
coded default name, a random default name is created every time the
script is run. This should prevent conflicts when running the install.sh
script again. The downside of this is that a new version of MIRACL will
be created every time the installation script is being run...