[feat/fix] add interactive data location prompt and bugfixes
If the user uses the interactive prompt, there will now be an option to
set the location for the data on the host system that will then
automatically be mounted at '/data' within the container. Also fixed
numerous smaller issues as well as inconsistencies caused by the new
interactive prompt.