[feat] implement inference and change interface folder object class
Implements inference script but it's a WIP. There are several variables
that get passed to the inference script that are not working yet. Mostly
this is because of type differences. I did improve the way subfolders
are handled as Pydantic objects. Each subfolder is a separate instance
of the interface subfolder class now.