[feat] add first, WIP implementation of MAPL3 workflow interface
The MAPL3 interface now returns a dictionary with the objects that are
required by the other modules in the workflow. It would probably make
sense to wrap the object generation in the main() fn of the MAPL3
interface into a class to be able to return all objects and then pick
and choose in the workflow instead of hardcoding the returned objects in
the dictionary. I'll make a note of this and implmement it during
refactoring once the time crunch is less and MAPL3 is stable for the
module and the workflow.