[feat/wip] first implementation of MAPL3 workflow
This is not working yet. The flags that are being used in the workflow
differ from the flags that are being used in the module. This is
intentional but when importing the MAPL3 module interface into the
workflow, the module interface expects the module flags/args to create
the objects with their content. However, the workflow args will be
parsed on the cli instead and the interface will throw an error because
an expected flag/args is missing.