[feat] major refactor of interface and processing methods for Pydantic objs
Ideally I would have liked to push this as several commits instead of
one big commit but time constraints won't allow that. This refactor was
needed to accomodate using the module as part of a encompassing
workflow. The interface can now be imported into the workflow and the
appropriate workflow flags will be used automatically. This will work
even better once the cli parsing of MIRACL has been refactored with
Typer.

Major changes:

1) The parser creator is now a class that consists of a method to create
the parser objects and a method to process the parsed cli arguments and
assign them automagically to their respective object/attr.

2) The arguments for the parser creation are now separated from the
parser scripts. I also added a dictionary with all Pydantic objects that
are used in the interface. This gives us a single source of truth that
can be imported/returned for processing. This is particularly useful for
creating workflows as we will have a separate workflow dictionary with
all objects needed for it.

3) The interface for the module is much shorter now with a lot less
duplication and variable assignment since the cli args are assigned
automatically. They attributes for the objects can now be referenced
directly using the objs dict.

4) All folders that are required for MAPL3's in-module workflow are now
created from a separate objects file. That way they can be handled
separately from the core MIRACL objects.

To do:

1) The interface should return two dicts, namely the MIRACL core objects
and MAPL3's folders.
2) The workflow has to be refactored to the same structure as the
module.
3) The dataclasses should be added to the ABC/abstractmethods in the
interface to avoid duplication.