[feat] first working version of MAPL3 workflow
A few things are still WIP but this is the first version of the MAPL3
workflow that actually works. The main additions here are the main
scripts for the MAPL3 module and workflows as well as re-defining the
input to the interfaces. This allows us to now call the MAPL3 module
interface with whatever parser/object combination that we want, in our
case either the module parser or the workflow parser.