[fix] fix incorrect GPU detection, change exit type to return and add helpful user msg
The script would fail because of the GPU check. The `nvidia-smi` query
would return multiple lines that would get parsed incorrectly, resulting
in the script to exit because no GPU's were detected. I changed the
query by counting the lines of the output with `| wc -l`. However, to
prevent this bug from happening again with different `nvidia-smi`
versions or changes in the future, the fn is not exiting the
installation script anymore when no GPU's are detected. Instead, the
entry for the GPU's is omitted from the `docker-compose.yml` and the
installation will proceed as normal. At the end, the user will be
informed that the GPU checks failed. If the they think that this is a
mistake, and they are confident that one or more Nvidia GPU's with CUDA
support are present, they are presented with the entry that they will
have to add in the `docker-compose.yml` and asked to do that manually.