[feat] add dowload directive for ACE models to install.sh
This might seem redundant since the models will at this point already be
downloaded through the download directive in the Dockerfile. However,
because we are mounting the 'miracl' folder of the local repo by
default, the 'models' folder will be overwritten by the local repo
folder. If the models are not downloaded to the local repo folder as
well, they will not be present when the local repo folder is mounted. If
the flag that disables mounting is set during installation, the models
will have been downloaded through the Dockerfile download directive and
will be available in the container.

It might be a good idea to forward the variables used in the install.sh
script for the models URLs to the Dockerfile to make sure that changing
one also changes the other. I will open an internal issue for that.