[feat] add additional flags for Docker build cache and verbose installation steps output
Docker build chache is enabled by default to guarantee clean builds when
rebuild from the same local repo. Setting the flag enables the chache.
The user can also choose to enable verbose output for the installation
steps. This is useful for debugging, especially when used with the '-l'
flag which enables writing a log file.