[feat] make not mounting miracl scripts folder less permanent
Instead of not adding the miracl scripts folder to the mounted volumes
when the corresponding flag is set, the folder is now still added but
commented out. The user will be informed that the folder can be mounted
by uncommenting the line. The if/else statement is a bit more robust now
as well.