[fix] fix breaking change in registration fn
Registering a nifti stack that was converted with a channel prefix and
number combination would break resampling tiff slices to original space.
The registration function did not have a check to account for prefix and
channel numbers i.e. all files in a data folder that contained e.g. both
signal and autofluorescent tiffs distinguished by a prefix/number would
be parsed. I added two new flags to the registration script: '-n' for
channel number and '-x' for channel prefix. They have to be supplied in
workflows as well.