[fix] change method to call Python script
Added a parser to the Python script that now expects positional args.
The reason the args are positional is because because I didnt' want
to have to duplicate flags between the Python and Bash scripts.
Obviously this will be changed once the Pydantic implementation is
finished. I also changed the method for calling the Python script
from within the Bash script to calling the script directly.