[fix] fix incorrect upsampling and downsampling between image and atlas spaces
The image now gets correctly upsampled to atlas space. Currently an
upsampling factor of 2 is hardcoded into the script. However, this will
be replaced with a user defined factor once the option is implemented as
a flag. After the segmentation has been created, it automatically gets
downsampled into image space. The reg_final folder now contains both the
image and atlas space segementations.

To make all of this work with the Waxholm reference atlas, we had to
skull strip it and clean it up and match the annotations to the
template. The updated annotation/template will have to be uploaded to
our HuggingFace repo as well.

We will also have to think about rat images that exceed the atlas space
dimensions and how we handle sampling in these cases.