[fix] fix incorrect olfactory bulb flag check
The logic to check whether the olfactory bulb should or shouldn't be
removed was incorrect. It would always evaluate to false since it would
only check for an empty flag or a None value. Adding a check for 0 fixed
the problem. Another issue is that the olfactory bulb templates are
hardcoded to 25vox. This needs to be fixed as well. A ticket has been
opened internally.