[feat] add functionality for differing orientations across subjects
This will close #150.

`miracl flow ace` has previously only accepted one orientation code for
all subjects. The analysis would be run incorrectly when some or all
subjects were oriented differently. To fix this, two new flags were
added to the workflow:

1) -rcaso,--rca_sep_orient_code

Set flag if orientations differ between individual subjects (default
i.e. flag not set: False). This requires an 'orientation.txt' file at
the root of the RAW Tiff data folder (i.e. in the folder where all RAW
Tiff images are stored) for EACH subject containing only the respective
orientation of the subject which must be exactly 3 uppercase letters
(e.g. 'ALS' or 'PLI'). This will override '--rca_orient_code' so make
sure that the 'orientation.txt' file exists in each subject's RAW Tiff
folder"

2) -rcaao,--rca_autodetect_sep_orient_code

Same as '--rca_sep_orient_code' except that it only expects an
'orientation.txt' file for subjects that have a different orientation
from the one provided to '--rca_orient_code` (default i.e. flag not set:
False). This will use the value provided to '--rca_orient_code` by
default unless an 'orientation.txt' file is detected in the RAW Tiff
folder of a subject, in which case the value from the 'orientation.txt'
file will be used. This is useful if you have many subjects with the
same orientation and only a few exceptions

The flags are mutually exclusive which is checked in the validation
method of the ACE workflow parser.

This commit also includes 7 unittests for the new functionality. While I
still have to figure out a better naming convention, this will hopefully
be the start of an improved test coverage for MIRACL.