[feat] custom runners for conv and reg
This is currently required as the argparsers for the conv and reg
modules are not yet integrated into the new registry architecture.
Instead, the workflow flags/values have to be manually remapped to the
module parsers. This can also be automated as all information is in the
objects. Eventually the default runner can be used once conv and reg
have been refactored away from hardcoded parser to object references.